Added test setup
