+++ auth/package.json
@@ -4,7 +4,15 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "ts-node-dev src/index.ts"
+    "start": "ts-node-dev src/index.ts",
+    "test": "jest --watchAll --no-cache"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node",
+    "setupFilesAfterEnv": [
+      "./src/test/setup.ts"
+    ]
   },
   "keywords": [],
   "author": "",
@@ -21,5 +29,13 @@
     "mongoose": "^5.12.13",
     "ts-node-dev": "^1.1.6",
     "typescript": "^4.3.2"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.23",
+    "@types/supertest": "^2.0.11",
+    "jest": "^27.0.4",
+    "mongodb-memory-server": "^6.9.6",
+    "supertest": "^6.1.3",
+    "ts-jest": "^27.0.3"
   }
 }
